Fix autotvm on windows

+++ python/tvm/autotvm/tuner/model_based_tuner.py
@@ -248,9 +248,9 @@ class ModelBasedTuner(Tuner):
             if self.trial_pt >= len(self.trials) - int(0.05 * self.plan_size):
                 # if the trial list is empty or
                 # the tuner is doing the last 5% trials (e-greedy), choose randomly
-                index = np.random.randint(len(self.space))
+                index = np.random.randint(len(self.space), dtype=np.int64)
                 while index in self.visited:
-                    index = np.random.randint(len(self.space))
+                    index = np.random.randint(len(self.space), dtype=np.int64)
 
             ret.append(self.space.get(index))
             self.visited.add(index)
